[Connector][WRS-3099] Split Acquia connector into multiple modules
